refactor: 버튼, 모달 페이지로 분리
